Move index.js to web_src and use webpack
