+++ .eslintrc.json
@@ -1,0 +1,5 @@
+{
+    "env": {
+        "es6": true
+    }
+}
